# Performance: Implement multi-tier caching architecture for improved cache efficiency